MOTECH-2088: moved from messages HeaderMapper and Transformer
