+++ platform/event/src/main/java/org/motechproject/event/messaging/impl/MotechEventTransformer.java
@@ -1,4 +1,4 @@
-package org.motechproject.event.messaging;
+package org.motechproject.event.messaging.impl;
 
 import org.motechproject.event.MotechEvent;
